refactor(http): [#184] move extractor to extractor mod
